Removed DeleteItem command/handler and tidied up exception classes
