+++ java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/api/v2/impl/QueryMenuControllerImplV2.java
@@ -2,7 +2,6 @@ package com.xxAMIDOxx.xxSTACKSxx.menu.api.v2.impl;
 
 import com.xxAMIDOxx.xxSTACKSxx.menu.api.v1.dto.response.MenuDTO;
 import com.xxAMIDOxx.xxSTACKSxx.menu.api.v2.QueryMenuControllerV2;
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
 import com.xxAMIDOxx.xxSTACKSxx.menu.commands.OperationCode;
 import com.xxAMIDOxx.xxSTACKSxx.menu.domain.Menu;
 import com.xxAMIDOxx.xxSTACKSxx.menu.exception.MenuNotFoundException;
@@ -30,9 +29,8 @@ public class QueryMenuControllerImplV2 implements QueryMenuControllerV2 {
         this.menuQueryService
             .findById(id)
             .orElseThrow(
-                () ->
-                    new MenuNotFoundException(
-                        new MenuCommand(OperationCode.GET_MENU_BY_ID, correlationId, id)));
+                () -> new MenuNotFoundException(id.toString(), OperationCode.GET_MENU_BY_ID.getCode(), correlationId));
+
     return ResponseEntity.ok(mapper.toMenuDto(menu));
   }
 }
