--- java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/commands/DeleteItemCommand.java
@@ -1,20 +1,0 @@
-package com.xxAMIDOxx.xxSTACKSxx.menu.commands;
-
-import java.util.UUID;
-import lombok.Getter;
-import lombok.Setter;
-
-/** @author ArathyKrishna */
-@Getter
-@Setter
-public class DeleteItemCommand extends MenuCommand {
-
-  private UUID categoryId;
-  private UUID itemId;
-
-  public DeleteItemCommand(String correlationId, UUID menuId, UUID categoryId, UUID itemId) {
-    super(OperationCode.DELETE_MENU_ITEM, correlationId, menuId);
-    this.categoryId = categoryId;
-    this.itemId = itemId;
-  }
-}
