--- java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/events/CategoryUpdatedEvent.java
@@ -1,11 +1,0 @@
-package com.xxAMIDOxx.xxSTACKSxx.menu.events;
-
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
-import java.util.UUID;
-
-public class CategoryUpdatedEvent extends CategoryEvent {
-
-  public CategoryUpdatedEvent(MenuCommand command, UUID categoryId) {
-    super(command, EventCode.CATEGORY_UPDATED, categoryId);
-  }
-}
