--- java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/events/MenuCreatedEvent.java
@@ -1,11 +1,0 @@
-package com.xxAMIDOxx.xxSTACKSxx.menu.events;
-
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
-import java.util.UUID;
-
-public class MenuCreatedEvent extends MenuEvent {
-
-  public MenuCreatedEvent(MenuCommand command, UUID menuId) {
-    super(command, EventCode.MENU_CREATED, menuId);
-  }
-}
