--- java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/events/MenuUpdatedEvent.java
@@ -1,10 +1,0 @@
-package com.xxAMIDOxx.xxSTACKSxx.menu.events;
-
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
-
-public class MenuUpdatedEvent extends MenuEvent {
-
-  public MenuUpdatedEvent(MenuCommand command) {
-    super(command, EventCode.MENU_UPDATED);
-  }
-}
