+++ java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/exception/MenuAlreadyExistsException.java
@@ -1,8 +1,5 @@
 package com.xxAMIDOxx.xxSTACKSxx.menu.exception;
 
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
-import java.util.UUID;
-
 public class MenuAlreadyExistsException extends MenuApiException {
 
   public MenuAlreadyExistsException(String restaurantId, String name, int operationCode, String correlationId) {
