+++ java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/menu/exception/MenuApiException.java
@@ -1,18 +1,9 @@
 package com.xxAMIDOxx.xxSTACKSxx.menu.exception;
 
 import com.xxAMIDOxx.xxSTACKSxx.core.api.exception.ApiException;
-import com.xxAMIDOxx.xxSTACKSxx.menu.commands.MenuCommand;
 import com.xxAMIDOxx.xxSTACKSxx.menu.commands.OperationCode;
 
 public class MenuApiException extends ApiException {
-
-  public MenuApiException(String message, ExceptionCode exceptionCode, MenuCommand menuCommand) {
-    super(
-        message,
-        exceptionCode,
-        OperationCode.fromCode(menuCommand.getOperationCode()),
-        menuCommand.getCorrelationId());
-  }
 
   public MenuApiException(String message, ExceptionCode exceptionCode, int operationCode, String correlationId) {
     super(
